fix: og and extract utility
